+++ app/src/main/java/com/example/spendsprout_opsc/categories/CategoryViewModel.kt
@@ -25,10 +25,10 @@ class CategoryViewModel {
                 val categories = BudgetApp.db.categoryDao().getAll().first()
                 val categoryList = mutableListOf<HierarchicalCategoryAdapter.CategoryWithSubcategories>()
                 for (category in categories) {
-                    val allSubcategories = getSubcategoriesForCategory(category.id.toLong())
+                    val allSubcategories = getSubcategoriesForCategory(category.id)
                     // Get only top 3 subcategories for overview
                     val top3Subcategories = allSubcategories.take(3)
-                    val categorySpent = getCategorySpent(category.id.toLong())
+                    val categorySpent = getCategorySpent(category.id)
                     val subcategorySpent = calculateTotalSpent(allSubcategories) // Use all for total calculation
                     val totalSpent = categorySpent + subcategorySpent
                     
@@ -69,7 +69,7 @@ class CategoryViewModel {
                 val categories = BudgetApp.db.categoryDao().getAll().first()
                 val categoryList = mutableListOf<Category>()
                 for (category in categories) {
-                    val spent = getCategorySpent(category.id.toLong())
+                    val spent = getCategorySpent(category.id)
                     categoryList.add(
                         Category(
                             id = category.id.toString(),
@@ -98,7 +98,7 @@ class CategoryViewModel {
                 val categories = BudgetApp.db.categoryDao().getAll().first()
                 val categoryList = mutableListOf<Category>()
                 for (category in categories) {
-                    val spent = getCategorySpentForDateRange(category.id.toLong(), startDate, endDate)
+                    val spent = getCategorySpentForDateRange(category.id, startDate, endDate)
                     categoryList.add(
                         Category(
                             id = category.id.toString(),
@@ -126,10 +126,10 @@ class CategoryViewModel {
         return "$sign R ${String.format("%.2f", absoluteAmount)}"
     }
     
-    private suspend fun getCategorySpent(categoryId: Long): Double {
+    private suspend fun getCategorySpent(categoryId: Int): Double {
         return try {
             val expenses = BudgetApp.db.expenseDao().getAll()
-            val categoryName = BudgetApp.db.categoryDao().getById(categoryId.toInt())?.categoryName
+            val categoryName = BudgetApp.db.categoryDao().getById(categoryId)?.categoryName
             if (categoryName != null) {
                 expenses.filter { it.expenseCategory == categoryName }
                     .sumOf { expense ->
@@ -148,10 +148,10 @@ class CategoryViewModel {
         }
     }
     
-    private suspend fun getCategorySpentForDateRange(categoryId: Long, startDate: Long, endDate: Long): Double {
+    private suspend fun getCategorySpentForDateRange(categoryId: Int, startDate: Long, endDate: Long): Double {
         return try {
             val expenses = BudgetApp.db.expenseDao().getAll()
-            val categoryName = BudgetApp.db.categoryDao().getById(categoryId.toInt())?.categoryName
+            val categoryName = BudgetApp.db.categoryDao().getById(categoryId)?.categoryName
             if (categoryName != null) {
                 expenses.filter { 
                     it.expenseCategory == categoryName && 
@@ -173,12 +173,12 @@ class CategoryViewModel {
         }
     }
     
-    private suspend fun getSubcategoriesForCategory(categoryId: Long): List<com.example.spendsprout_opsc.wants.model.Subcategory> {
+    private suspend fun getSubcategoriesForCategory(categoryId: Int): List<com.example.spendsprout_opsc.wants.model.Subcategory> {
         return try {
             val subcategoryEntities = BudgetApp.db.subcategoryDao().getByCategoryId(categoryId)
             subcategoryEntities.map { subcategory ->
                 // Calculate actual spent amount from transactions for this subcategory
-                val actualSpent = getSubcategorySpent(subcategory.id.toLong())
+                val actualSpent = getSubcategorySpent(subcategory.id)
                 com.example.spendsprout_opsc.wants.model.Subcategory(
                     id = subcategory.id.toString(),
                     name = subcategory.subcategoryName,
@@ -192,10 +192,10 @@ class CategoryViewModel {
         }
     }
     
-    private suspend fun getSubcategorySpent(subcategoryId: Long): Double {
+    private suspend fun getSubcategorySpent(subcategoryId: Int): Double {
         return try {
             val expenses = BudgetApp.db.expenseDao().getAll()
-            val subcategoryName = BudgetApp.db.subcategoryDao().getById(subcategoryId.toInt())?.subcategoryName
+            val subcategoryName = BudgetApp.db.subcategoryDao().getById(subcategoryId)?.subcategoryName
             if (subcategoryName != null) {
                 expenses.filter { it.expenseCategory == subcategoryName }
                     .sumOf { expense ->
@@ -293,7 +293,7 @@ class CategoryViewModel {
                 val category = categories.find { it.categoryName.equals(categoryName, ignoreCase = true) }
                 
                 if (category != null) {
-                    val subcategories = getSubcategoriesForCategory(category.id.toLong())
+                    val subcategories = getSubcategoriesForCategory(category.id)
                     CoroutineScope(Dispatchers.Main).launch {
                         callback(subcategories)
                     }
